Retrace: properly abort when frameretrace_server dies
Previously, the UI would spin at 100% when the server crashed.
Instead, provide a message to the user letting them know what
happened.

Pivotal-tracker: https://www.pivotaltracker.com/story/show/154485573
Requested-by: Arthur Huillet <arthur.huillet@free.fr>

+++ retrace/daemon/glframe_socket.cpp
@@ -82,10 +82,10 @@ Socket::Write(const void * buf, int size) {
   int bytes_remaining = size;
   const void *curPtr = buf;
   while (bytes_remaining > 0) {
-    size_t bytes_written = ::send(m_socket_fd,
-                                  reinterpret_cast<const char *>(curPtr),
-                                  bytes_remaining,
-                                  0);  // default flags
+    int bytes_written = ::send(m_socket_fd,
+                               reinterpret_cast<const char *>(curPtr),
+                               bytes_remaining,
+                               0);  // default flags
 
     if (bytes_written < 0) {
       if (errno == EINTR)
